+++ src/main/java/br/com/zalf/prolog/webservice/interno/suporte/SuporteService.java
@@ -9,6 +9,7 @@ import br.com.zalf.prolog.webservice.commons.util.StringUtils;
 import br.com.zalf.prolog.webservice.interno.PrologInternalUser;
 import br.com.zalf.prolog.webservice.interno.autenticacao.AutenticacaoInternaService;
 import br.com.zalf.prolog.webservice.interno.suporte._model.InternalEmpresa;
+import br.com.zalf.prolog.webservice.interno.suporte._model.InternalUnidade;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.io.FilenameUtils;
 import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
@@ -80,6 +81,34 @@ public final class SuporteService {
             Log.e(TAG, String.format("Erro ao alterar imagem da empresa %d", codEmpresa), throwable);
             throw new RuntimeException(throwable);
         }
+    }
+
+    @NotNull
+    public Response insertUnidade(@NotNull final String authorization,
+                                  @NotNull final InternalUnidade unidade) {
+        validate(authorization);
+        dao.insertUnidade(unidade);
+        return Response.ok("Unidade alterada com sucesso!");
+    }
+
+    @NotNull
+    public List<InternalUnidade> getTodasUnidades(@NotNull final String authorization) {
+        validate(authorization);
+        return dao.getTodasUnidades();
+    }
+
+    @NotNull
+    public InternalUnidade getUnidade(@NotNull final String authorization,
+                                      @NotNull final Long codUnidade) {
+        validate(authorization);
+        return dao.getUnidade(codUnidade);
+    }
+
+    @NotNull
+    public Response updateUnidade(final String authorization, @NotNull final InternalUnidade unidade) {
+        final PrologInternalUser user = validate(authorization);
+        dao.updateUnidade(unidade, user);
+        return Response.ok("Unidade alterada com sucesso!");
     }
 
     @NotNull