+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Change Log
 ## UNRELEASED
 ##
 
+### Features
+* Cria crud de unidades para uso no intralog
+
 
 <a name="v3.7.0"></a>
 ## Version [v3.7.0](https://github.com/luizfp/PrologWebService/compare/v3.6.0...v3.7.0) (2021-08-04)